Fixed missing behavior / bug where recursive functions don't work defined in the scope of another function.
For example:

```
let wrapper = fn() {
  let inner = fn(x) {
    if (x == 0) {
      return 0;
    } else {
      return inner(x - 1);
    }
  };
  return inner(1);
};

wrapper();
``

This is fixed by implicitly putting the current closure into the first
"free" variable slot with `SetSelf` opcode.

A side effect of this is that it would mean shadowing the nema of the
function wouldn't be possible, so this diff also fixes that as well by
allowing `self` here (*meaning the current closure's bound name*) to be
shadowed by a local binding. This therefore also allows:

```
let one = fn() {
  let one = 1
  return one
}
one()
``